+++ libX11/src/xcms/cmsColNm.c
@@ -126,13 +126,13 @@ _XcmsColorSpaceOfString(
  */
 {
     XcmsColorSpace	**papColorSpaces;
-    int n;
+    size_t n;
     char *pchar;
 
     if ((pchar = strchr(color_string, ':')) == NULL) {
 	return(XcmsFailure);
     }
-    n = (int)(pchar - color_string);
+    n = (size_t)(pchar - color_string);
 
     if (ccc == NULL) {
 	return(NULL);
@@ -198,7 +198,7 @@ _XcmsParseColorString(
     XcmsColorSpace	*pColorSpace;
     char		string_buf[64];
     char		*string_lowered;
-    int			len;
+    size_t		len;
     int			res;
 
     if (ccc == NULL) {
